[SQL] Break dataTypes.scala into multiple files.
It was over 1000 lines of code, making it harder to find all the types.
